feat: 更新SPI和定时器驱动 - 优化SPI中断处理逻辑 - 优化事件处理和数据传输逻辑 - 调整定时器PWM配置

+++ Drivers/STM32F10x_StdPeriph/drv_serial.c
@@ -241,7 +241,7 @@ static bool _check_uart_name_valid(const char *name)
     const char *valid_names[] = {"UART1",  "UART2",  "UART3",
                                  "USART1", "USART2", "USART3"};
 
-    for (size_t i = 0; i < sizeof(valid_names) / sizeof(valid_names[0]); i++)
+    for (uint16_t i = 0; i < sizeof(valid_names) / sizeof(valid_names[0]); i++)
     {
         if (_strcasecmp_upper(name, valid_names[i]))
         {
